% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrir_apresentacao.R
\name{abrir_apresentacao}
\alias{abrir_apresentacao}
\title{Abrir apresentação do curso}
\usage{
abrir_apresentacao()
}
\value{
A apresentação é aberta em formato PDF no navegador
}
\description{
Abre a apresentação do curso
}
\examples{
abrir_apresentacao()
}
\author{
Alexandre Jaloto
}
