% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrir_livro.R
\name{abrir_livro}
\alias{abrir_livro}
\title{Abrir livro do curso}
\usage{
abrir_livro()
}
\value{
O livro é aberto no navegador
}
\description{
Abre livro com os comandos do curso
}
\examples{
abrir_livro()
}
\author{
Alexandre Jaloto
}
