% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banco.2pl.R
\docType{data}
\name{banco.2pl}
\alias{banco.2pl}
\title{Banco simulado de itens dicotômicos}
\format{
An object of class \code{data.frame} with 500 rows and 11 columns.
}
\usage{
banco.2pl
}
\description{
Banco simulado de itens dicotômicos
}
\details{
Banco simulado com 500 sujeitos e 10 itens dicotômicos, em que 1
equivale à resposta correta e 0, à errada. A variável \code{grupo} indica o
grupo a que cada sujeito pertence. A simulação prevê DIF no item 2
entre os grupos.
}
\keyword{datasets}
