% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banco.poli.R
\docType{data}
\name{banco.poli}
\alias{banco.poli}
\title{Banco simulado de itens politômicos}
\format{
An object of class \code{data.frame} with 500 rows and 10 columns.
}
\usage{
banco.poli
}
\description{
Banco simulado de itens politômicos
}
\details{
Banco simulado com 500 sujeitos e 10 itens politômicos, em que 3
equivale à resposta mais alta e 0, à mais baixa.
}
\keyword{datasets}
