% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrir_apresentacao.R
\name{abrir_apresentacao}
\alias{abrir_apresentacao}
\title{Abrir apresentação do curso}
\usage{
abrir_apresentacao()
}
\value{
O arquivo é aberto no programa adequado.
}
\description{
Abre arquivo PPTX com a apresentação do curso
}
\examples{
abrir_apresentacao()
}
\author{
Alexandre Jaloto
}
